+++ src/migrations/Install.php
@@ -21,7 +21,7 @@ class Install extends Migration
 		$this->createTable(
 			CookieConsent::SITE_SETTINGS_TABLE,
 			[
-				'site_id' => $this->integer(11),
+				'site_id' => $this->primaryKey(),
 				'dateCreated' => $this->dateTime()->notNull(),
 				'dateUpdated' => $this->dateTime()->notNull(),
 				'uid'         => $this->uid(),
@@ -39,11 +39,6 @@ class Install extends Migration
                 'refresh' => $this->boolean()->notNull()->defaultValue(false),
                 'refresh_time' => $this->integer()->notNull()->defaultValue(500)
 			]
-		);
-		$this->addPrimaryKey(
-			'pk_cookie_consent_site_settings',
-			CookieConsent::SITE_SETTINGS_TABLE,
-			'site_id'
 		);
 		$this->addForeignKey(
 			'fk_cookie_consent_setting_belong_to_site',
@@ -95,12 +90,8 @@ class Install extends Migration
                 'cookieName' => $this->string()->defaultValue('cookie-consent'),
 				'dateCreated' => $this->dateTime()->notNull(),
 				'dateUpdated' => $this->dateTime()->notNull(),
-			]
-		);
-		$this->addPrimaryKey(
-			'pk_cookie_consent_consent',
-			CookieConsent::CONSENT_TABLE,
-			'uid'
+                'PRIMARY KEY(uid)',
+			],
 		);
 		$this->addForeignKey(
 			'fk_cookie_consent_consent_belong_to_site',